Changes middlewares and apply the changes in app

+++ part4_exercises/blog_app/controllers/blogs.js
@@ -11,15 +11,9 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!request.token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
-  }
+  const user = request.user
 
-
-  const user = await User.findById(decodedToken.id)
-
-  console.log(body)
+  console.log(user)
   const blog = new Blog({
     title: body.title,
     author: body.author,
@@ -48,8 +42,17 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if(user._id.equals(blog.user)){
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
+  else {
+    response.status(404).json({ error: 'error with deleted blog' })
+  }
 
 })
